ASoC: arizona_marley: fref must be limited in pseudo-fractional mode
When the FLL is in pseudo-fractional mode there is an additional
limit on fref based on the fratio, to prevent aliasing around the
Nyquist frequency. If fref exceeds this limit the refclk divider
must be increased and the calculation tried again until a suitable
combination of fref and fratio is found or we have to fall back to
integer mode.

This patch also adds some debug log prints around this code.

Change-Id: Ia6a17b847423c7908ec2b83436f37ea31e52a0c4
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>